Fix a bug that caused global variables initialized in Buildfiles to not be visible to class definitions inside the same file
This is due to a quirk in how Python does variable lookup while in exec: initializing a variable in exec puts it in locals(), while accessing the variable looks in globals(). This is because Python assumes they are the same thing.

Also, CLEAN now also deletes the final build artifact, not only the intermediate object files.