Add missing import
I only caught this because I used pylsp, which uses pyflakes internally.

The reason this worked so far is that utils imports sys, so utils.sys is a valid object. Therefore, when I do from utils import *, it add utils.sys to the global namespace, which is exactly the same as the sys module.